+++ src/OptionParser_v2.cpp
@@ -512,22 +512,55 @@ std::string serializeResult(const ParseResult &result) {
     return output_string;
 }
 
-void generateUsageStringImpl(std::stringstream &ss,
-                             const Component &component) {
-    ss << "[";
+void generateUsageStringImpl(std::stringstream &ss, const Component &component,
+                             bool skip_square_brackets = false) {
+    if (!component.isRequired() && !skip_square_brackets) {
+        ss << "[";
+    }
+
     if (component.isParameter()) {
-        ss << "<" << component.getName() << "> ";
+        ss << "<" << component.getName() << ">";
     } else if (component.isFlag()) {
-        ss << component.getName() << " ";
+        ss << component.getName();
     } else if (component.isCommand()) {
-        ss << component.getName() << " ";
+        ss << component.getName();
     } else {
         assert(false);
     }
-    for (const auto &child : component.getChildren()) {
-        generateUsageStringImpl(ss, child.get());
+
+    const auto children = component.getChildren();
+    int n_child = children.size();
+    if (!children.empty()) {
+        ss << " ";
     }
-    ss << "]";
+
+    // flags
+    for (const auto &flag : component.getFlags()) {
+        generateUsageStringImpl(ss, flag.get());
+        if (--n_child > 0) {
+            ss << " ";
+        }
+    }
+
+    // commands
+    if (!component.getCommands().empty()) {
+        ss << "<command>";
+        if (--n_child > 0) {
+            ss << " ";
+        }
+    }
+
+    // parameters
+    for (const auto &parameter : component.getParameters()) {
+        generateUsageStringImpl(ss, parameter.get());
+        if (--n_child > 0) {
+            ss << " ";
+        }
+    }
+
+    if (!component.isRequired() && !skip_square_brackets) {
+        ss << "]";
+    }
 }
 
 std::string generateUsageString(const Component &root_component) {
@@ -535,32 +568,50 @@ std::string generateUsageString(const Component &root_component) {
 
     ss << "Usage: ";
 
-    if (root_component.isParameter()) {
-        ss << "<" << root_component.getName() << "> ";
-    } else {
-        ss << root_component.getName() << " ";
-    }
-
-    // flags
-    for (const auto &flag : root_component.getFlags()) {
-        generateUsageStringImpl(ss, flag.get());
-    }
-
-    // commands
-    for (const auto &command : root_component.getCommands()) {
-        generateUsageStringImpl(ss, command.get());
-    }
-
-    // parameters
-    for (const auto &parameter : root_component.getParameters()) {
-        generateUsageStringImpl(ss, parameter.get());
-    }
+    generateUsageStringImpl(ss, root_component, true);
 
     return ss.str();
 }
 
-std::string generateHelpString(const Component &root_component) {
+void generateHelpStringImpl(std::stringstream &ss, const Component &component,
+                            int indent = 0, int margin = 40) {
+    // write root component and description
+    ss << std::string(indent, ' ') << component.getName();
+
+    // write parameters
+    int parameter_text_size = 0;
+    for (const auto &parameter : component.getParameters()) {
+        ss << " <" << parameter.get().getName() << ">";
+        parameter_text_size += parameter.get().getName().size() + 3;
+    }
+
+    // calculate the remaining space for description alignment
+    int remainingSpace = margin - (indent + component.getName().size() + 1 +
+                                   parameter_text_size);
+
+    if (remainingSpace > 0) {
+        ss << std::string(remainingSpace, ' ') << " ";
+    } else {
+        ss << std::endl << std::string(margin, ' ');
+    }
+
+    ss << component.getDescription() << std::endl;
+
+    // write flags
+    for (const auto &flag : component.getFlags()) {
+        generateHelpStringImpl(ss, flag.get(), indent + 4, margin);
+    }
+
+    // write commands
+    for (const auto &command : component.getCommands()) {
+        generateHelpStringImpl(ss, command.get(), indent + 4, margin);
+    }
+}
+
+std::string generateHelpString(const Component &root_component, int margin) {
     std::stringstream ss;
+
+    generateHelpStringImpl(ss, root_component, 0, margin);
 
     return ss.str();
 }